Add test coverage for data exporting
Runs the exports but intercepts the zip file without loading into S3 so the contents can be inspected and assertions made.
