+++ lib/utils/snake-case.js
@@ -1,0 +1,1 @@
+module.exports = str => str.replace(/[A-Z]/g, s => `_${s.toLowerCase()}`);
