+++ package.json
@@ -5,7 +5,11 @@
   "main": "index.js",
   "scripts": {
     "dev": "nodemon -r dotenv/config index.js",
-    "test": "eslint ."
+    "test:lint": "eslint .",
+    "pretest:unit": "npm run migrate",
+    "test:unit": "mocha ./test/specs --recursive",
+    "migrate": "DATABASE_NAME=asl-test NODE_ENV=test knex migrate:latest --knexfile ./node_modules/@asl/schema/knexfile.js",
+    "test": "npm run test:lint && npm run test:unit"
   },
   "repository": {
     "type": "git",
@@ -22,10 +26,15 @@
   },
   "homepage": "https://github.com/UKHomeOffice/asl-data-exports#readme",
   "devDependencies": {
+    "bl": "^5.0.0",
+    "csv-parse": "^4.16.3",
     "dotenv": "^8.2.0",
     "eslint": "^7.23.0",
     "eslint-config-lennym": "^2.2.0",
-    "nodemon": "^2.0.7"
+    "jszip": "^3.7.1",
+    "mocha": "^9.1.1",
+    "nodemon": "^2.0.7",
+    "uuid": "^8.3.2"
   },
   "dependencies": {
     "@asl/schema": "^10.4.1",
